Updated Tables

+++ IspDataDb/dbo/Tables/Holiday.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Holiday] (
     [IdNo]          SMALLINT   IDENTITY (1, 1) NOT NULL,
-    [LeaveIdNo]     INT        NULL,
+    [LeaveIdNo]     SMALLINT   NULL,
     [DateStart]     DATE       NULL,
     [DateEnd]       DATE       NULL,
     [EnteredBy]     INT        NULL,
@@ -8,6 +8,8 @@
     [DateTimeStamp] ROWVERSION NULL,
     CONSTRAINT [PK_Holiday] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
